api: add an option to skip hook if no resource is found
A boolean field called SkipHookIfNotPresent is added to the Hook type.
This boolean has the default value of false.

If the field is set to true then on timeout of the hook, the hook
operation is considered as a success if no resources were found matching
the selectors provided.

If one or more resources are found but the condition cannot be satisfied
or the operation cannot be completed then the hook operation is
considered as a failure.

Signed-off-by: Raghavendra Talur <raghavendra.talur@gmail.com>